Upgrade to Python3
Ensemble.py was changed so that the pd0-parsing code now works with
Python3. Previously the parsing code only worked in Python2 due to
the use of the buffer() function.

In Python3, the built-in buffer() function was replaced with the
built-in memoryview() function. However, instead of using the
memoryview() function, the following struct function is used instead: struct.unpack_from(format-string, buffer, offset)

The memoryview() function may be used in the future to improved
memory usage and increase the speed of execution.