Build ensembles as individual numpy arrays
Before these changes, ensembles were built by incrementally growing
a pandas dataframe. As it turns out, this was a huge slow down.
Now, individual ensembles are compressed into a single long numpy
array that is pre-allocated in size (once fixed leader is parsed,
the exact size of the array is known). Then, the ensemble arrays
are collected into a python list in the time series object. Once
all ensembles are collected, ensemble list is converted into a
pandas dataframe.

The nominal size of an ensemble is: 721
  - 26 values from variable leader
  - 4*40 = 160 velocity values
  - 4*40 = 160 correlation values
  - 4*40 = 160 echo intensity values
  - 4*40 = 160 percent good values
  - 55 bottom track values

Fixed Leader information is stored in Ensemble and TimeSeries
attributes (instead of as part of the pandas dataframe).

To try and limited highly repeated code, `parse_velocity` was
replaced with `parse_water_profiling_data`.