Improve documentation and introduce plotting functions
In the README file, table of contents was created with linking
between sections. Also, units and variable information was
added for variables stored in PathfinderEnsemble and
PathfinderTimeSeries. Note that units listed in the documentation
convey metric-converted units, not native .pd0 file units.

The other update from this commit is the introduction of plotting
functionality via Jupyter Notebook interface. Note that the notebook
plotting functions are still a bit unorganized, but this will be
improved in future commits. By taking all of the plotting methods
out of the main Pathfinder files, the vehicle will not be burdened
with unnecessary plotting overhead during deployments. The 'figs/'
directory is used to store outputs of generated plots.