Create skeleton for parent class for DVL objects
Parent class for DVL is made to improved the overall organization of
the repository. This will also make it easier to access and edit data
in the DVL data arrays and combine ensembles together into a
time-series.

NOTE: still significant coding and testing required to validate that
superclass works with DVL parsing.

Also, this allows for easy definition of "short" and "long" variable
lists, where the short version is a compressed version that is ideal
for onboard DVL navigation, and the long version refers to saving all
possible data fields (which would create unnecessary weight to onboard
navigation process). The short version is 70% decrease from the full
data from an ensemble.

Additionally, the following updates were made:
  - increased organization of DVL figures (added directory structure)
  - moved data out of the git repository and stored in Dropbox instead