Extract slope and orientation from DEM map
Provided a Digital Elevation Map (DEM), this commit computes the slope
and orientation of the seafloor surface using the convolution with the
Scharr kernel.

The map is first resized, using the haversine to extract map extent
information, such that the pixel width and height cover the same
distance. This will make future processing easier.