Correct for mounting bias, pitch and roll bias
Use rotation matrices to account for pitch and roll bias terms.
Heading bias has not yet been determined. After implementing
propagation methods for ocean currents, the GPS fix will be compared
to the DVL odometry result, and any discrepancies can be partially
attributed to a bias in heading.

Now, vertical speed due to change in pressure is not computed when
there is no change in pressure reading. This is done because the depth
sensor reports values asynchronously from the vehicle compared to the
DVL instrument.

Multiple bins are considered when extracting through-water velocity
estimates from the DVL. It is observed that the bin directly below the
transducer head tends to under-estimate the through-water velocity.