Shear propagation method on last Sentinel mission
Shear propagation method applied to last sentinel dive in the dataset
and strong performance is observed. Though I believe that the there
is more tuning to be done to make sure that the approach extends well
to other dives. For this, some tuning parameters have been selected.

This commit contains the tuning parameters that yield 1.5 percent
of distance traveled for localization error when applied to Sentinel
last mission.