Implements simplified AUG simulator
Simulator indicates range statistics, energy consumption, and
endurance. Simulator applied several control policies to a
Strawman mission in the Arctic.

Note that most of the code changes consist of spaghetti code in
the form of Jupyter notebooks. This commit is made to preserve a
working copy of the code before a proper code cleanup is done.