Implement three-factor extraction from bathymetry
Converted Kolumbo bathymetry map from Lat/Lon coordinates to UTM
coordinates. Then, the map the map is saved using Pandas DataFrame
for easing plotting alongside AUG odometry. (created `BathymetryMap`
for this)

Created `AdaptiveVelocityController` class to perform energy-optimal
propulsion control, according to the methods outlined in the Frontiers
paper (still need to add EDBS to this file).

Implemented three-factor extraction within PathfinderEnsemble to parse
the three seafloor factors whenever they are available.

In the process of cleaning up the Jupyter notebook for this project,
but there is still a lot of work here.