Adaptive-TAU, Batch-Testing, varying bathymetric resolutions, sensitvity testing
MF-TAN: Put hooks into Point Cloud class that allows for the Grid Resolution to be passed in as an attribute and therefore change the Minimum Spatial resolution for a point cloud before feature extraction. Changed

MIN_SPATIAL RESOLUTION = 0.5 * Grid Resolution (instead of 2 * grid resolution)

Batch Testing File --> allows for sensitivity analysis for different heuristics and improved in-depth analysis

bathy-plotting pt 2 --> Can generate bathymetry charts at varying resolutions (for both the map and the "variance map"

In General: the nav algorithm was improved to select TAU values for MF-TAN based on the variability of the map prior for the area that the vehicle thinks its in.